# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "DistributionShiftResponse": {
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "evidence_state": {
            "title": "Evidence State",
            "type": "string"
          },
          "explanation_only": {
            "default": true,
            "title": "Explanation Only",
            "type": "boolean"
          },
          "feature_version": {
            "title": "Feature Version",
            "type": "string"
          },
          "previous_as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous As Of"
          },
          "skew_shift": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skew Shift"
          },
          "strategy_effect": {
            "default": false,
            "title": "Strategy Effect",
            "type": "boolean"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "tail_mass_change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tail Mass Change"
          },
          "tenors": {
            "items": {
              "type": "integer"
            },
            "title": "Tenors",
            "type": "array"
          },
          "term_shift": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term Shift"
          },
          "w1_shift": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "W1 Shift"
          }
        },
        "required": [
          "symbol",
          "as_of",
          "feature_version",
          "tenors",
          "evidence_state",
          "details"
        ],
        "title": "DistributionShiftResponse",
        "type": "object"
      },
...
      "EventStudyResponse": {
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "event_kind": {
            "title": "Event Kind",
            "type": "string"
          },
          "evidence_state": {
            "title": "Evidence State",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/EventStudyRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "event_kind",
          "as_of",
          "evidence_state",
          "rows"
        ],
        "title": "EventStudyResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptionsHistoryToggleInput": {
        "properties": {
          "lock_version": {
            "title": "Lock Version",
            "type": "integer"
          },
          "requested_state": {
            "enum": [
              "on",
              "off"
            ],
            "title": "Requested State",
            "type": "string"
          }
        },
        "required": [
          "requested_state",
          "lock_version"
        ],
        "title": "OptionsHistoryToggleInput",
        "type": "object"
      },
...
      "PortfolioTransactionInput": {
        "properties": {
          "account": {
            "default": "manual",
            "title": "Account",
            "type": "string"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "executed_at": {
            "title": "Executed At",
            "type": "string"
          },
          "expected_position_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Position Version"
          },
          "fees": {
            "default": 0,
            "title": "Fees",
            "type": "number"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "transaction_type": {
            "enum": [
              "opening_balance",
              "buy",
              "sell",
              "dividend",
              "fee",
              "split",
              "transfer_in",
              "transfer_out"
            ],
            "title": "Transaction Type",
            "type": "string"
          }
        },
        "required": [
          "transaction_type",
          "executed_at",
          "idempotency_key"
        ],
        "title": "PortfolioTransactionInput",
        "type": "object"
      },
      "PortfolioTransactionPreviewResponse": {
        "additionalProperties": true,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "fees": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fees"
          },
          "new_average_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Average Cost"
          },
          "new_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Quantity"
          },
          "old_average_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Average Cost"
          },
          "old_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Quantity"
          },
          "position_version": {
            "title": "Position Version",
            "type": "string"
          },
          "realized_pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Pnl"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "transaction_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Type"
          }
        },
        "required": [
          "position_version"
        ],
        "title": "PortfolioTransactionPreviewResponse",
        "type": "object"
      },
      "PortfolioTransactionResultResponse": {
        "properties": {
          "portfolio": {
            "$ref": "#/components/schemas/TablePayloadResponse"
          },
          "transaction": {
            "additionalProperties": true,
            "title": "Transaction",
            "type": "object"
          }
        },
        "required": [
          "transaction",
          "portfolio"
        ],
        "title": "PortfolioTransactionResultResponse",
        "type": "object"
      },
      "PortfolioTransactionReversalInput": {
        "properties": {
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key"
        ],
        "title": "PortfolioTransactionReversalInput",
        "type": "object"
      },
...
      "SourceDetailResponse": {
        "additionalProperties": true,
        "properties": {
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          }
        },
        "title": "SourceDetailResponse",
        "type": "object"
      },
...
      "SuperinvestorDetailResponse": {
        "additionalProperties": true,
        "properties": {
          "holdings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Holdings",
            "type": "array"
          },
          "investor_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investor Key"
          }
        },
        "title": "SuperinvestorDetailResponse",
        "type": "object"
      },
      "TablePayloadResponse": {
        "properties": {
          "count": {
            "default": 0,
            "title": "Count",
            "type": "integer"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiStatusResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "TablePayloadResponse",
        "type": "object"
      },
...
      "WatchlistMutationResponse": {
        "properties": {
          "data_refresh": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Refresh"
          },
          "watchlist": {
            "$ref": "#/components/schemas/TablePayloadResponse"
          },
          "watchlist_symbol": {
            "additionalProperties": true,
            "title": "Watchlist Symbol",
            "type": "object"
          }
        },
        "required": [
          "watchlist_symbol",
          "watchlist"
        ],
        "title": "WatchlistMutationResponse",
        "type": "object"
      },
      "WatchlistSymbolInput": {
        "properties": {
          "asset_class": {
            "default": "equity",
            "title": "Asset Class",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "WatchlistSymbolInput",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Personal Investment Panel",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/options/event-study": {
      "get": {
        "operationId": "option_event_study_api_options_event_study_get",
        "parameters": [
          {
            "in": "query",
            "name": "ticker",
            "required": true,
            "schema": {
              "maxLength": 16,
              "minLength": 1,
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "event_kind",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Event Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "As Of",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStudyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Option Event Study"
      }
    },
...
    "/api/options/history/distribution-shift": {
      "get": {
        "operationId": "option_distribution_shift_api_options_history_distribution_shift_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "default": "QQQ",
              "maxLength": 16,
              "minLength": 1,
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "As Of",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionShiftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Option Distribution Shift"
      }
    },
...
    "/api/portfolio/transactions": {
      "get": {
        "operationId": "portfolio_transactions_api_portfolio_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablePayloadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Portfolio Transactions"
      },
      "post": {
        "operationId": "create_portfolio_transaction_api_portfolio_transactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioTransactionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioTransactionResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Portfolio Transaction"
      }
    },
    "/api/portfolio/transactions/preview": {
      "post": {
        "operationId": "preview_transaction_api_portfolio_transactions_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioTransactionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioTransactionPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Transaction"
      }
    },
    "/api/portfolio/transactions/{transaction_id}/reverse": {
      "post": {
        "operationId": "reverse_transaction_api_portfolio_transactions__transaction_id__reverse_post",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioTransactionReversalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioTransactionResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reverse Transaction"
      }
    },
...
    "/api/sources/{source_id}": {
      "get": {
        "operationId": "source_detail_api_sources__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Detail"
      }
    },
...
    "/api/superinvestors/{investor_key}": {
      "get": {
        "operationId": "superinvestor_detail_api_superinvestors__investor_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "investor_key",
            "required": true,
            "schema": {
              "title": "Investor Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuperinvestorDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Superinvestor Detail"
      }
    },
...
    "/api/watchlist/symbols": {
      "post": {
        "operationId": "save_watchlist_symbol_endpoint_api_watchlist_symbols_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistSymbolInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Watchlist Symbol Endpoint"
      }
    },
    "/api/watchlist/symbols/{symbol}": {
      "delete": {
        "operationId": "delete_watchlist_symbol_endpoint_api_watchlist_symbols__symbol__delete",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watchlist Symbol Endpoint"
      }
    },
    "/api/watchlist/symbols/{symbol}/options-history": {
      "patch": {
        "operationId": "set_watchlist_options_history_endpoint_api_watchlist_symbols__symbol__options_history_patch",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptionsHistoryToggleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsHistoryPolicyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Watchlist Options History Endpoint"
      }
    }
  }
}